+++ cpp/Makefile
@@ -41,7 +41,7 @@ LDFLAGS = -Wl,--gc-sections -pie
 LIBS = -pthread
 
 SRCS = md5.c sha1.c sha256.c kcapi-kernel-if.c kcapi-md.c kcapi-utils.c
-SRCS_CPP = diskhasher.cpp filehash.cpp hashclass.cpp
+SRCS_CPP = diskhasher.cpp filehash.cpp hash.cpp
 
 OBJS = $(SRCS:.c=.o)
 OBJS_CPP = $(SRCS_CPP:.cpp=.o)
