+++ cpp/include/hash.h
